Refactor MainWindow to support multiple scenes
