+++ src/main/resources/view/Homepage.fxml
@@ -1,0 +1,47 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<?import javafx.scene.layout.VBox?>
+
+<?import javafx.scene.control.SplitPane?>
+<?import javafx.geometry.Insets?>
+<?import javafx.scene.layout.StackPane?>
+
+<VBox xmlns="http://javafx.com/javafx/10.0.1" xmlns:fx="http://javafx.com/fxml/1">
+    <SplitPane id="testMSplitPane" fx:id="testMSplitPane" dividerPositions="0.5" VBox.vgrow="ALWAYS">
+        <VBox fx:id="testMQuestionListView" minWidth="340" prefWidth="340" SplitPane.resizableWithParent="false">
+            <padding>
+                <Insets top="10" right="10" bottom="10" left="10" />
+            </padding>
+            <StackPane fx:id="testMQuestionListPanelPlaceholder" VBox.vgrow="ALWAYS"/>
+        </VBox>
+
+        <VBox fx:id="testMAnswerListView" minWidth="340" prefWidth="340" SplitPane.resizableWithParent="false">
+            <padding>
+                <Insets top="10" right="10" bottom="10" left="10" />
+            </padding>
+            <StackPane fx:id="testMAnswerListPanelPlaceholder" VBox.vgrow="ALWAYS"/>
+        </VBox>
+    </SplitPane>
+
+    <SplitPane id="splitPane" fx:id="splitPane" dividerPositions="0.4" VBox.vgrow="ALWAYS">
+        <VBox fx:id="personList" minWidth="340" prefWidth="340" SplitPane.resizableWithParent="false">
+            <padding>
+                <Insets top="10" right="10" bottom="10" left="10" />
+            </padding>
+            <StackPane fx:id="personListPanelPlaceholder" VBox.vgrow="ALWAYS"/>
+        </VBox>
+
+        <VBox fx:id="cardList" minWidth="340" prefWidth="340" SplitPane.resizableWithParent="false">
+            <padding>
+                <Insets top="10" right="10" bottom="10" left="10" />
+            </padding>
+            <StackPane fx:id="cardListPanelPlaceholder" VBox.vgrow="ALWAYS"/>
+        </VBox>
+
+        <StackPane fx:id="browserPlaceholder" prefWidth="340" >
+            <padding>
+                <Insets top="10" right="10" bottom="10" left="10" />
+            </padding>
+        </StackPane>
+    </SplitPane>
+</VBox>
